* co_id_t, instance, device unit, debug level and some helper functions:   Changed from unsigned long to unsigned int.
git-svn-id: svn://svn.code.sf.net/p/colinux/code/branches/devel@1176 688f6386-cf32-0410-9ead-a596f0484e40

+++ src/colinux/common/common_base.h
@@ -127,7 +127,7 @@ typedef long co_rc_t;
 /*
  * Defines a LINUX instance. There are CO_MAX_COLINUXS of these 
  */
-typedef unsigned long co_id_t;
+typedef unsigned int co_id_t;
 
 #define CO_INVALID_ID ((co_id_t)-1)
